refactor: extract the asterisk painter into a shared util
