+++ lib/src/widgets/flow_thinking_indicator.dart
@@ -1,8 +1,7 @@
-import 'dart:math' as math;
-
 import 'package:material_ui/material_ui.dart';
 
 import '../theme/flow_theme.dart';
+import '../utils/flow_asterisk_painter.dart';
 import 'flow_shimmer_text.dart';
 
 /// The thinking line shown while the assistant has not started responding:
@@ -158,7 +157,7 @@ class _FlowThinkingIndicatorState extends State<FlowThinkingIndicator>
             opacity: depth.drive(Tween<double>(begin: 1, end: _minOpacity)),
             child: CustomPaint(
               size: Size.square(widget.size),
-              painter: _AsteriskPainter(
+              painter: FlowAsteriskPainter(
                 color: widget.color ?? colors.onSurfaceMuted,
                 // Proportional so the mark keeps its weight at any size.
                 strokeWidth: widget.size / 9,
@@ -195,48 +194,4 @@ class _FlowThinkingIndicatorState extends State<FlowThinkingIndicator>
       ],
     );
   }
-}
-
-/// The six-armed asterisk: three rounded strokes through the center. Drawn
-/// rather than shipped — no SDK glyph matches the design's mark, and paint
-/// stays crisp at any size and tint.
-class _AsteriskPainter extends CustomPainter {
-  const _AsteriskPainter({required this.color, required this.strokeWidth});
-
-  final Color color;
-  final double strokeWidth;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = color
-      ..strokeWidth = strokeWidth
-      ..strokeCap = StrokeCap.round
-      ..style = PaintingStyle.stroke;
-    // A translucent ink would composite twice where the strokes cross,
-    // darkening the hub against text in the same ink. Flatten the mark
-    // into one layer and apply the ink's alpha to the whole glyph once.
-    final translucent = color.a < 1;
-    if (translucent) {
-      canvas.saveLayer(
-        (Offset.zero & size).inflate(strokeWidth),
-        Paint()..color = const Color(0xFFFFFFFF).withValues(alpha: color.a),
-      );
-      paint.color = color.withValues(alpha: 1);
-    }
-    final center = size.center(Offset.zero);
-    final radius = (size.shortestSide - strokeWidth) / 2;
-    for (var i = 0; i < 3; i++) {
-      // Three diameters at 60° steps, starting upright so the resting mark
-      // has the design's vertical arm.
-      final angle = math.pi / 2 + math.pi * i / 3;
-      final delta = Offset(math.cos(angle), math.sin(angle)) * radius;
-      canvas.drawLine(center - delta, center + delta, paint);
-    }
-    if (translucent) canvas.restore();
-  }
-
-  @override
-  bool shouldRepaint(_AsteriskPainter oldDelegate) =>
-      oldDelegate.color != color || oldDelegate.strokeWidth != strokeWidth;
 }
